day22:
refine delete case, make it saver

+++ app/Http/Controllers/Admin/CaseManagement/CaseController.php
@@ -22,11 +22,14 @@ use App\WorkHistory;
 class CaseController extends Controller
 {
     //
-    public function create() {
+    public function create()
+    {
         return view('case.create');
     }
 
-    public function store(CreateCaseFormRequest $request) {
+    public function store(CreateCaseFormRequest $request)
+    {
+        $currentUser = User::where('email', $request->creator)->first();
         $input = $request->only('email', 'first_name', 'last_name', 'birthday', 'gender', 'webpage', 'ssn', 'ilp', 'ethnicity', 'program');
         $case = new CreateCase;
         $case->email = $request->get('email');
@@ -39,11 +42,14 @@ class CaseController extends Controller
         $case->ilp = $request->get('ilp');
         $case->ethnicity = $request->get('ethnicity');
         $case->program = $request->get('program');
+        $case->cm_id = $currentUser->id;
+        $case->cm_name = $currentUser->last_name . ', ' . $currentUser->first_name;
         $case->save();
         return redirect('admin/case/create')->withFlashMessage('Case Successfully Created and Activated!');
     }
 
-    public function view() {
+    public function view()
+    {
 //        $data = CreateCase::all()->sortByDesc('id')->paginate(10);
         //orderBy('id','desc')
         $data = CreateCase::orderBy('id', 'desc')->paginate(10);
@@ -51,7 +57,9 @@ class CaseController extends Controller
             'datas' => $data,
         ]);
     }
-    public function viewdetail($id) {
+
+    public function viewdetail($id)
+    {
         $data = CreateCase::find($id);
         $email = $data->email;
         $caseUser = User::where('email', $email)->first();
@@ -64,13 +72,17 @@ class CaseController extends Controller
             'workhistorys' => $workhistorys,
         ]);
     }
-    public function editdetail($id) {
+
+    public function editdetail($id)
+    {
         $data = CreateCase::find($id);
         return view('case.edit', [
             'data' => $data,
         ]);
     }
-    public function update($id, UpdateCaseFormRequest $request) {
+
+    public function update($id, UpdateCaseFormRequest $request)
+    {
         $case = CreateCase::find($id);
         $case->first_name = $request->get('first_name');
         $case->last_name = $request->get('last_name');
@@ -82,39 +94,52 @@ class CaseController extends Controller
         $case->ethnicity = $request->get('ethnicity');
         $case->program = $request->get('program');
         $case->save();
-        return redirect('admin/case/'. $id . '/view');
+        return redirect('admin/case/' . $id . '/view');
     }
-    public function test() {
+
+    public function test()
+    {
         return view('case.edit');
     }
 
-    public static function active($id) {
+    public static function active($id)
+    {
         $case = CreateCase::find($id);
         $case->status = 1;
         $case->save();
-        return redirect('/admin/case/'.$id.'/view');
+        return redirect('/admin/case/' . $id . '/view');
     }
 
-    public static function inactive($id) {
+    public static function inactive($id)
+    {
         $case = CreateCase::find($id);
         $case->status = 0;
         $case->save();
-        return redirect('/admin/case/'.$id.'/view');
+        return redirect('/admin/case/' . $id . '/view');
     }
 
-    public function delete($id) {
-        CreateCase::find($id)->delete();
-        return redirect('/admin/case/view');
+    public function delete($id, Request $request)
+    {
+        $case = CreateCase::find($id);
+        $name = $case->last_name.', '.$case->first_name;
+        if ($request->youth_name == $name) {
+            CreateCase::find($id)->delete();
+            return redirect('/admin/case/view');
+        } else {
+            return redirect()->back();
+        }
     }
 
-    public function createaccount($id) {
+    public function createaccount($id)
+    {
         $case = CreateCase::find($id);
         return view('case.account', [
             'case' => $case,
         ]);
     }
 
-    public function storeaccount($id, Request $request) {
+    public function storeaccount($id, Request $request)
+    {
         $case = CreateCase::find($id);
         $first_name = $case->first_name;
         $last_name = $case->last_name;
@@ -151,7 +176,8 @@ class CaseController extends Controller
         $newprofile->save();
     }
 
-    public function editfile(Request $request) {
+    public function editfile(Request $request)
+    {
         $case_id = $request->get('id');
         $doc_id = $request->get('doc_id');
         $doc = Docs::find($doc_id);
@@ -160,16 +186,20 @@ class CaseController extends Controller
         $doc->type = $request->get('type');
         $doc->description = $request->get('description');
         $doc->save();
-        return redirect('admin/case/'.$case_id.'/view');
+        return redirect('admin/case/' . $case_id . '/view');
     }
-    public function deletefile($id) {
+
+    public function deletefile($id)
+    {
         $doc = Docs::find($id);
-        $doc_path_name = $doc->path.'/'.$doc->filename;
+        $doc_path_name = $doc->path . '/' . $doc->filename;
         File::delete($doc_path_name);
         $doc->delete();
         return redirect()->back();
     }
-    public function storeWorkHistory(Request $request) {
+
+    public function storeWorkHistory(Request $request)
+    {
         $workhistory = new WorkHistory;
         $workhistory->case_id = $request->get('id');
         $workhistory->start_date = $request->get('start_date');
@@ -181,7 +211,9 @@ class CaseController extends Controller
         $workhistory->save();
         return redirect()->back();
     }
-    public function editWorkHistory($id, Request $request) {
+
+    public function editWorkHistory($id, Request $request)
+    {
         $case_id = $request->get('id');
         $workhistoryid = $id;
         $workhistory = WorkHistory::find($workhistoryid);
@@ -194,7 +226,9 @@ class CaseController extends Controller
         $workhistory->save();
         return redirect()->back();
     }
-    public function deleteWorkHistory($id) {
+
+    public function deleteWorkHistory($id)
+    {
         $workhistory = WorkHistory::find($id);
         $workhistory->delete();
         return redirect()->back();
